MPI projects and example solutions

+++ content/courses/parallel_computing_introduction/codes/random_walk.cxx
@@ -16,7 +16,7 @@ int main(int argc, char **argv) {
     mt19937 rng(rd());
     uniform_int_distribution<int> choice(1,4);
 
-    int N;
+    long long N;
     if (argc != 2) {
         cout<<"0,0,0\n";
         return 1;
@@ -28,24 +28,24 @@ int main(int argc, char **argv) {
         ssteps>>N;
     }
 
-    int x=0;
-    int y=0;
+    double x=0;
+    double y=0;
     int direction;
 
-    for (int i=1; i<=N; ++i) {
+    for (long long i=1; i<=N; ++i) {
         direction=choice(rng);
         switch (direction) {
             case 1:
-                x+=1;
+                x+=1.;
                 break;
             case 2:
-                x-=1;
+                x-=1.;
                 break;
             case 3:
-                y+=1;
+                y+=1.;
                 break;
             case 4:
-                y-=1;
+                y-=1.;
                 break;
         }
     }
