+++ content/courses/parallel_computing_introduction/solns/mpirandom_walk.cxx
@@ -23,7 +23,7 @@ int main(int argc, char **argv) {
    MPI_Comm_size(MPI_COMM_WORLD, &npes);
    MPI_Comm_rank(MPI_COMM_WORLD, &rank);
 
-   int N;
+   long long N;
 
    if (argc != 2) {
       cout<<rank<<":0,0,0\n";
@@ -37,25 +37,25 @@ int main(int argc, char **argv) {
       ssteps>>N;
    }
 
-   int x=0;
-   int y=0;
+   double x=0.;
+   double y=0.;
 
    int direction;
 
-   for (int i=1; i<=N; ++i) {
+   for (long long i=1; i<=N; ++i) {
       direction=choice(rng);
       switch (direction) {
          case 1: 
-           x+=1;
+           x+=1.;
            break;
          case 2: 
-           x-=1;
+           x-=1.;
            break;
          case 3: 
-           y+=1;
+           y+=1.;
            break;
          case 4: 
-           y-=1;
+           y-=1.;
            break;
       }
    }
